reduced dashboard componetns to memoItem and LocationItem
